refactor: depr instructor tabs
